More requested changes
Made DamageType an item field to allow conversion of vanilla items, changed SetX to ModifyX, etc
