+++ patches/tModLoader/Terraria/ModLoader/DamageClassLoader.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 
 namespace Terraria.ModLoader
 {
@@ -11,13 +10,10 @@ namespace Terraria.ModLoader
 
 		internal static bool DamageClassExists(ModDamageClass damageClass) => damageClass == null ? false : damageClasses.Contains(damageClass);
 
-		internal static void Add(ModDamageClass damageClass) {
+		internal static int Add(ModDamageClass damageClass) {
 			damageClasses.Add(damageClass);
+			return damageClasses.Count - 1;
 		}
-
-		internal static int GetIndex<T>() => damageClasses.FindIndex(d => d.GetType() == typeof(T));
-
-		internal static int GetIndex(ModDamageClass damageClass) => damageClasses.FindIndex(d => d == damageClass);
 
 		internal static void Unload() {
 			damageClasses.Clear();
