+++ patches/tModLoader/Terraria/ModLoader/ModDamageClass.cs
@@ -2,6 +2,8 @@
 {
 	public abstract class ModDamageClass : ModType
 	{
+		internal int index;
+
 		/// <summary>
 		/// This is the name that will show up when an item tooltip displays 'X [name] damage'.
 		/// </summary>
@@ -9,7 +11,7 @@
 
 		protected override void Register() {
 			Mod.damageClasses[Name] = this;
-			DamageClassLoader.Add(this);
+			index = DamageClassLoader.Add(this);
 			ContentInstance.Register(this);
 		}
 	}
