+++ patches/tModLoader/Terraria/Player_tML.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using Terraria.ModLoader;
+﻿using Terraria.ModLoader;
 
 namespace Terraria
 {
@@ -18,81 +17,48 @@ namespace Terraria
 		/// Edits the crit for the given damage type on this player.
 		/// </summary>
 		/// <param name="changeAmount">The amount that should be added to the crit. Can be negative too.</param>
-		public void SetCrit<T>(int changeAmount) where T : ModDamageClass {
-			int damageClassType = DamageClassLoader.GetIndex<T>();
-			if (ContainsIndex(damageData, damageClassType)) {
-				ref var data = ref damageData[damageClassType];
-				data.crit += changeAmount;
-			}
+		public void ModifyCrit<T>(int changeAmount) where T : ModDamageClass {
+			ref var data = ref damageData[ModContent.GetInstance<T>().index];
+			data.crit += changeAmount;
 		}
 
 		/// <summary>
 		/// Edits the damage value for the given damage type on this player.
 		/// </summary>
 		/// <param name="changeAmount">The amount (as a percentage, 0.01f is 1%) that should be added to the damage. Can be negative too.</param>
-		public void SetDamage<T>(float changeAmount) where T : ModDamageClass {
-			int damageClassType = DamageClassLoader.GetIndex<T>();
-			if (ContainsIndex(damageData, damageClassType)) {
-				ref var data = ref damageData[damageClassType];
-				data.damage += changeAmount;
-			}
+		public void ModifyDamage<T>(float changeAmount) where T : ModDamageClass {
+			ref var data = ref damageData[ModContent.GetInstance<T>().index];
+			data.damage += changeAmount;
 		}
 
 		/// <summary>
 		/// Edits the damage multiplier's value for the given damage type on this player.
 		/// </summary>
 		/// <param name="changeAmount">The amount (as a percentage, 0.01f is 1%) that should be added to the damage multiplier. Can be negative too.</param>
-		public void SetDamageMult<T>(float changeAmount) where T : ModDamageClass {
-			int damageClassType = DamageClassLoader.GetIndex<T>();
-			if (ContainsIndex(damageData, damageClassType)) {
-				ref var data = ref damageData[damageClassType];
-				data.damageMult += changeAmount;
-			}
+		public void ModifyDamageMult<T>(float changeAmount) where T : ModDamageClass {
+			ref var data = ref damageData[ModContent.GetInstance<T>().index];
+			data.damageMult += changeAmount;
 		}
 
 		/// <summary>
 		/// Gets the crit stat for this damage type on this player.
 		/// </summary>
-		public int GetCrit<T>() where T : ModDamageClass {
-			int damageClassType = DamageClassLoader.GetIndex<T>();
-			return GetCrit(damageClassType);
-		}
+		public int GetCrit<T>() where T : ModDamageClass => GetCrit(ModContent.GetInstance<T>().index);
 
 		/// <summary>
 		/// Gets the damage stat for this damage type on this player.
 		/// </summary>
-		public float GetDamage<T>() where T : ModDamageClass {
-			int damageClassType = DamageClassLoader.GetIndex<T>();
-			return GetDamage(damageClassType);
-		}
+		public float GetDamage<T>() where T : ModDamageClass => GetDamage(ModContent.GetInstance<T>().index);
 
 		/// <summary>
 		/// Gets the damage multiplier stat for this damage type on this player.
 		/// </summary>
-		public float GetDamageMult<T>() where T : ModDamageClass {
-			int damageClassType = DamageClassLoader.GetIndex<T>();
-			return GetDamageMult(damageClassType);
-		}
+		public float GetDamageMult<T>() where T : ModDamageClass => GetDamageMult(ModContent.GetInstance<T>().index);
 
 		internal int GetCrit(int damageClassType) => damageData[damageClassType].crit;
 		
 		internal float GetDamage(int damageClassType) => damageData[damageClassType].damage;
 		
 		internal float GetDamageMult(int damageClassType) => damageData[damageClassType].damageMult;
-
-		private bool ContainsIndex(Array array, int index) => index >= 0 && index <= array.Length - 1;
-	}
-
-	public struct DamageClassData
-	{
-		public int crit;
-		public float damage;
-		public float damageMult;
-
-		public DamageClassData(int crit, float damage, float damageMult) {
-			this.crit = crit;
-			this.damage = damage;
-			this.damageMult = damageMult;
-		}
 	}
 }
